requirements.txt
autogen-agentchat>=0.4.0
autogen-ext[openai,azure]>=0.4.0
python-dotenv>=1.0.1
azure-search-documents
tqdm
autogen_core
streamlit
